fast forward funcion

+++ src/instructions.h
@@ -21,7 +21,7 @@ struct emulator_state
 {
     uint8_t fetched; //Recently fetched data for current operation
     uint16_t addr_abs; //Absolute Address for current instruction
-    uint8_t addr_rel; //Relative address for branching
+    uint16_t addr_rel; //Relative address for branching
     uint8_t opcode; //Current opcode being executed
 };
